First cut at the new broken RPS game

+++ code/zips.sh
@@ -1,9 +1,11 @@
 #! /bin/sh
 
 for f in *; do
-    if [[ -d $f ]]; then
-        echo $f is a directory
-	rm $f.zip
-	zip -r $f.zip $f
+    if [ "$f" != "old" ] ; then
+        if [[ -d $f ]]; then
+            echo $f is a directory
+	    rm $f.zip
+	    zip -r $f.zip $f
+        fi
     fi
 done
